backend/requirements.txt
# Backend Dependencies for Garuda Electricals & Hardwares
# FastAPI Application

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
supabase==2.4.0
python-multipart==0.0.6
reportlab==4.0.8
email-validator==2.1.0.post1
httpx>=0.24.0
